+++ testdata/xliff/goldens/tsout/zh_CN.ts
@@ -2,7 +2,7 @@
     // Do not modify this file by hand!
     // Re-generate this file by running lit-localize
 
-    
+    import {html} from 'lit-html';
 
     /* eslint-disable no-irregular-whitespace */
     /* eslint-disable @typescript-eslint/camelcase */