+++ CHANGELOG.md
@@ -5,7 +5,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## Unreleased
+<!-- ## Unreleased -->
+
+## [0.2.0] - 2020-05-13
 
 - Add support for the XLIFF localization interchange format:
   https://docs.oasis-open.org/xliff/v1.2/os/xliff-core.html